Fix: next.js로 마이그레이션
