--- FrontendWeb/src/Router.tsx
@@ -1,10 +1,0 @@
-import { Route, Routes } from 'react-router-dom';
-import Home from './pages/Home';
-
-export default function Router() {
-  return (
-    <Routes>
-      <Route path="/" element={<Home />} />
-    </Routes>
-  );
-}
