+++ frontend-web/.prettierrc.json
@@ -1,0 +1,8 @@
+{
+    "printWidth": 120,
+    "tabWidth": 4,
+    "singleQuote": true,
+    "trailingComma": "all",
+    "semi": true,
+    "plugins": ["prettier-plugin-tailwindcss"]
+}
