+++ frontend-web/src/app/layout.tsx
@@ -1,0 +1,15 @@
+import StyledComponentsRegistry from '@/lib/registry'
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <html>
+      <body>
+        <StyledComponentsRegistry>{children}</StyledComponentsRegistry>
+      </body>
+    </html>
+  )
+}
