+++ frontend-web/src/components/Button/FormSubmitButton/index.tsx
@@ -1,0 +1,18 @@
+import Button from '..';
+import { FormSubmitButtonProps } from '../../../types/button';
+
+export default function FormSubmitButton(props: FormSubmitButtonProps) {
+  return (
+    <Button
+      width="24px"
+      height="auto"
+      backgroundSize="auto"
+      {...props}
+      position={'absolute'}
+      zIndex={1}
+      background='none'
+    >
+      {props.children}
+    </Button>
+  );
+}
